VideoIndexerDemo/VideoIndexer/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
exception==0.1.0
Flask==1.1.1
Flask-Cors==3.0.7
flask-restx==0.2.0
idna==2.9
importlib-metadata==1.7.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
msgpack==0.6.1
nltk==3.5.0
numpy==1.19.0
Pillow==7.1.2
pyrsistent==0.16.0
pysrt==1.1.2
python-dateutil==2.8.1
python-docx==0.8.10
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
Shapely==1.7.0
six==1.15.0
tqdm==4.46.1
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
